nifstd migrate to orthauth

+++ nifstd/test/test_integration.py
@@ -1,7 +1,7 @@
 import unittest
 from pathlib import Path
 from pyontutils.utils import get_working_dir
-from pyontutils.config import devconfig
+from pyontutils.config import auth
 from pyontutils.integration_test_helper import _TestScriptsBase, Folders, Repo
 import nifstd_tools
 
@@ -37,8 +37,11 @@ if working_dir is None:
     # a number of problems with references to local vs installed packages
     working_dir = module_parent
 
-devconfig._check_ontology_local_repo()  # FIXME maybe we can be a bit less blunt about it?
-ont_repo = Repo(devconfig.ontology_local_repo)
+olr = auth.get_path('ontology-local-repo')
+if not olr.exists():
+    raise FileNotFoundError(f'{olr} does not exist cannot continue')
+
+ont_repo = Repo(olr)
 post_load = lambda : (ont_repo.remove_diff_untracked(), ont_repo.checkout_diff_tracked())
 post_main = lambda : (ont_repo.remove_diff_untracked(), ont_repo.checkout_diff_tracked())
